Rendering Stuff Done (read desc)
Changed Iridescent's renderer functions to glib2d. Because the font renderer was messed up on the PSP. It's practically the same but we think there was missing calls with the initialization. So I changed calls to glib2d and deleted sprite texture stuff because glib2d has a image renderer aswell. Other than that everything is working right now :).